enter info to direct jump
jumps to section in chapter 5.3.3 the hub. coco note to jump to another page in the manual, enter the exact name of the specific section you want to jump to in the manual and put brackets around it... do this twice and see the colors change from black to red and blue. voila!