rm duplicate content
removed evaluation projects section that was originally inserted to be a place to add details on measures and metrics for the eval projects. The 21st century content was duplicated in Specific Projects, so the 21st content was pasted in specific projects for review. CSP and Camp Rio contents were duplicate or incomplete and removed from chapter 4. TSLIP sentences were moved to specific projects.